+++ src/RookieTools/__init__.py
@@ -5,11 +5,14 @@ Author: Rookie
 E-mail: hyll8882019@outlook.com
 """
 from RookieTools.common import *
-from RookieTools.PasswordCheckBase import *
+from RookieTools.PasswordCheckBase import PassWordCheckBase
+from RookieTools.WebFileCheckBase import WebFileCheckBase
+from RookieTools.ip import SpecialIp, is_special_ip
+from RookieTools.logger import logger
 
-__version__ = '1.11'
+__version__ = '1.15'
 
 __all__ = [
     'PassWordCheckBase', 'show_run_time', 'downloader', 'logger', 'get_charset', 'html_decode', 'is_open_port',
-    'host2ip', 'is_url'
+    'host2ip', 'is_url', 'is_special_ip', 'SpecialIp', 'pool', 'common', 'WebFileCheckBase'
 ]